DPKW-1-编写设备轮询worker 1.写入新文件
